skip mutation when beyond conc. limit of assay
Added elif statement to read_fep_data function on lines 66/67:
    This statement determines that a particular mutation (L248R/imatinib or T315I/dasatinib) have IC50s that are beyond the concentration limits of the assay (vis. IC50 is 10000 nM).

Same deal for reading in the Prime data on lines 369/370, except here the input data file is just two columns AND we deal with the experimental values on free energy scales (5.08 kcal/mol for T315I/dasatinib and 2.33 kcal/mol for L248R/imatinib). These floats are unique in the data file, so this elif logic works as intended.